+++ quantify_scheduler/backends/circuit_to_device.py
@@ -3,9 +3,12 @@
 """
 Compilation backend for quantum-circuit to quantum-device layer.
 """
+import warnings
 from itertools import permutations
 from copy import deepcopy
 from typing import Any, Callable, Dict, List, Optional, Union
+import numpy as np
+
 
 from quantify_scheduler.helpers.importers import import_python_object_from_string
 from quantify_scheduler.operations.operation import Operation
@@ -107,7 +110,8 @@ class DeviceCompilationConfig(DataStructure):
 
 
 def compile_circuit_to_device(
-    schedule: Schedule, device_cfg: Union[DeviceCompilationConfig, dict, None]
+    schedule: Schedule,
+    device_cfg: Optional[Union[DeviceCompilationConfig, dict]] = None,
 ) -> Schedule:
     """
     Adds the information required to represent operations on the quantum-device
@@ -142,11 +146,7 @@ def compile_circuit_to_device(
     for operation in schedule.operations.values():
         # if operation is a valid pulse or acquisition it will not attempt to add
         # pulse/acquisition info in the lines below.
-        if operation.valid_pulse:
-            _verify_pulse_clock_present(operation=operation, schedule=schedule)
-            continue
-        if operation.valid_acquisition:
-            # no verification at this point.
+        if operation.valid_pulse or operation.valid_acquisition:
             continue
 
         qubits = operation.data["gate_info"]["qubits"]
@@ -180,15 +180,106 @@ def compile_circuit_to_device(
                 device_cfg=device_cfg,
             )
 
-        pulse_acq_info = operation["pulse_info"] + operation["acquisition_info"]
-        clocks_used = []
-        for info in pulse_acq_info:
-            clocks_used.append(info["clock"])
+    return schedule
 
+
+def set_pulse_and_acquisition_clock(
+    schedule: Schedule,
+    device_cfg: Optional[Union[DeviceCompilationConfig, dict]] = None,
+) -> Schedule:
+    """
+    Ensures that each pulse/acquisition-level clock resource has either been added
+    to the schedule or, if present in device_cfg, adds it to the schedule.
+    A warning is given when a clock resource has conflicting frequency
+    definitions, and an error is raised if the clock resource is unknown.
+
+    Parameters
+    ----------
+    schedule
+        The schedule to be compiled.
+    device_cfg
+        Device specific configuration, defines the compilation step from
+        the quantum-circuit layer to the quantum-device layer description.
+        Note, if a dictionary is passed, it will be parsed to a
+        :class:`~DeviceCompilationConfig`.
+
+    Warns
+    -----
+    RuntimeWarning
+        When clock has conflicting frequency definitions.
+
+    Raises
+    ------
+    RunTimeError
+        When operation is not at pulse/acquisition-level.
+    ValueError
+        When clock frequency is unknown.
+    ValueError
+        When clock frequency is NaN.
+    """
+    if not isinstance(device_cfg, DeviceCompilationConfig):
+        # this is a special case to be supported to enable compilation for schedules
+        # that are defined completely at the quantum-device layer and require no
+        # circuit to device compilation.
+        # A better solution would be to omit skip this compile call in a backend,
+        # but this is supported for backwards compatibility reasons.
+        if device_cfg is not None:
+            device_cfg = DeviceCompilationConfig.parse_obj(device_cfg)
+
+    # to prevent the original input schedule from being modified.
+    schedule = deepcopy(schedule)
+
+    for operation in schedule.operations.values():
+        # if the operation is at gate-level it must be compiled from
+        # circuit to device first.
+        if not operation.valid_pulse and not operation.valid_acquisition:
+            raise RuntimeError(
+                f"Operation '{operation}' is a gate-level operation and must be "
+                f"compiled from circuit to device; ensure compilation "
+                f"is made in the correct order."
+            )
+
+        clocks_used = []
+        for info in operation["pulse_info"] + operation["acquisition_info"]:
+            clocks_used.append(info["clock"])
         for clock in set(clocks_used):
-            if clock not in schedule.resources:
-                frequency = device_cfg.clocks[clock]
-                clock_resource = ClockResource(name=clock, freq=frequency)
+            # if clock is defined both in the schedule and device_cfg,
+            # ensures the frequency is the same in both
+            if clock in schedule.resources:
+                if device_cfg is not None and clock in device_cfg.clocks:
+                    clock_freq_device_cfg = device_cfg.clocks[clock]
+                    clock_freq_schedule = schedule.resources[clock]["freq"]
+                    if (
+                        not np.isnan(clock_freq_device_cfg)
+                        and clock_freq_device_cfg != clock_freq_schedule
+                    ):
+                        warnings.warn(
+                            f"Clock '{clock}' has conflicting frequency definitions: "
+                            f"{clock_freq_schedule} Hz in the schedule and "
+                            f"{clock_freq_device_cfg} Hz in the device config. "
+                            f"The clock is set to '{clock_freq_schedule}'. "
+                            f"Ensure the schedule clock resource matches the "
+                            f"device config clock frequency or set the "
+                            f"clock frequency in the device config to np.NaN "
+                            f"to omit this warning.",
+                            RuntimeWarning,
+                        )
+            else:
+                if device_cfg is None or clock not in device_cfg.clocks:
+                    raise ValueError(
+                        f"Operation '{operation}' contains an unknown clock '{clock}'; "
+                        f"ensure this resource has been added to the schedule "
+                        f"or to the device config."
+                    )
+
+                if np.isnan(clock_freq_device_cfg := device_cfg.clocks[clock]):
+                    raise ValueError(
+                        f"Operation '{operation}' contains clock '{clock}' with an "
+                        f"undefined frequency; ensure this resource has been added to "
+                        f"the schedule or to the device config."
+                    )
+
+                clock_resource = ClockResource(name=clock, freq=clock_freq_device_cfg)
                 schedule.add_resource(clock_resource)
 
     return schedule
@@ -299,18 +390,6 @@ def _add_device_repr_from_cfg(
 
     device_op = factory_func(**factory_kwargs)
     operation.add_device_representation(device_op)
-
-
-def _verify_pulse_clock_present(operation, schedule):
-    for pulse in operation["pulse_info"]:
-        if "clock" in pulse:
-            if pulse["clock"] not in schedule.resources:
-                raise ValueError(
-                    "Operation '{}' contains an unknown clock '{}'; ensure "
-                    "this resource has been added to the schedule.".format(
-                        str(operation), pulse["clock"]
-                    )
-                )
 
 
 def _add_device_repr_from_cfg_multiplexed(