+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Deprecation - Refactored tests to remove deprecated `qcompile`, refactored to `SerialCompiler` (!529, #368)
 - Qblox backend - Add preparation of acquisition settings and accompanying datastructures for NV centers (!567)
 - Waveforms - Fix `sudden_net_zero` waveform generation function misunderstands `amp_B` (!549, #390)
+- Compilation - Fix clock not being added to schedule from quantum device layer via new graph node `set_pulse_and_acquisition_clock` that verifies and sets clock frequency (!538,#371)
 - Schedule Functions - make experiment-related schedule functions available in `quantify_scheduler.schedules` (!572)
 - Tests - Removed unused `lo0` and added `ttl_acq_threshold` in `qblox_test_mapping_nv_centers.json` so that `TestNVDarkESRSched` suite passes (!579)
 - Documentation - Removed deprecated code from the Compiling to Hardware Tutorial (!582)